Created new setting in plugin settings in admin section: Update Interval.
Update interval is the interval to which the script that pulls in fresh
data from the yahoo finance api is ran.

Added html to OptionsPage for setting

Added additional function in processSettings to handle processing the form
and saving setting to the database.

Created new yahoo classes that hold the code to actually run the update
and store the results (YahooApiResults).

Saved the sql that creates the table that will hold all pulled in data
from yahoo (trade_date, trade_time, trade_price).

Updated the test for YahooFinanceAPI class.